third_party/WebKit/LayoutTests/bluetooth/characteristic/notifications/characteristic-does-not-support-notifications.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  // The temperature type characteristic only supports 'read'.
  return getTemperatureTypeCharacteristic()
    .then(({characteristic}) => assert_promise_rejects_with_message(
      characteristic.startNotifications(),
      new DOMException('GATT Error: Not supported.',
                       'NotSupportedError'),
      'Characteristic does not support notifications.'));
}, 'Characteristic does not support notifications. Reject with ' +
   'NotSupportedError.');
</script>
